Improved 'Complete Installation' steps
