+++ src/app/components/installations/complete-installation/select-installation/SelectInstallation.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Table, TableBody, TableHeader, TableHeaderColumn, TableRow, TableRowColumn, RaisedButton, } from 'material-ui';
-import { IP } from '../../../../config/config.js';
-import { dateHelpers } from '../helpers/common.js';
+import { IP } from '../../../../../../config/config.js';
+import { dateHelpers } from '../../../helpers/common.js';
 
 export default class SelectInstallation extends React.Component {
 
@@ -9,35 +9,15 @@ export default class SelectInstallation extends React.Component {
     super(props);
 
     this.state = {
-      //an array of all scheduled installations
-      installations: undefined,
-
       //selected installation
       selectedInstallation: undefined,
     };
   }
 
-  componentDidMount() {
-    var httpRequest = new XMLHttpRequest();
-    let _this = this;
-    httpRequest.onreadystatechange = function() {
-      if (this.readyState == 4 && this.status == 200) {
-        let installations = JSON.parse(httpRequest.responseText).installations;
-        _this.setState({
-          installations: installations
-        });
-
-      }
-    };
-
-    httpRequest.open('GET', "http://" + IP + "/installations/getallscheduled?id=" + this.props.userId, true);
-    httpRequest.send(null);
-  }
-
   handleRowSelected(selectedRows) {
     if(selectedRows.length == 1) {
       this.setState({
-        selectedInstallation: this.state.installations[selectedRows[0]],
+        selectedInstallation: this.props.installations[selectedRows[0]],
         selectedNum: selectedRows[0],
       });
     } else {
@@ -87,7 +67,7 @@ export default class SelectInstallation extends React.Component {
             </TableRow>
           </TableHeader>
           <TableBody deselectOnClickaway={false}>
-            {this.state.installations? this.state.installations.map( (row, index) => (
+            {this.props.installations? this.props.installations.map( (row, index) => (
                 <TableRow key={index} selected={index == this.state.selectedNum ? true : false}>
                   <TableRowColumn>{row.customerName}</TableRowColumn>
                   <TableRowColumn>{row.address}</TableRowColumn>
